+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Let users specify a filepath during an auto-mock. This filepath will be read
   from disk when an auto-mock is used.
 - Improve `nextjs-graphql-example`. Add a demo of an image placeholder.
+- Add Chrome Extension notifications. Config updates fire notifications.
 
 ## 0.0.9 NextJS demo, better watch mode